feat/#62 : JWT를 가공하는 JwtResolver 클래스 생성
JwtManager 책임 분리
